containers/linux-mcp-server/requirements.txt
mcp[cli]
fastapi
uvicorn
sse-starlette
psutil
python-dotenv
asyncssh
